Cleanup 'signing' into 2 packages: sypgp, signing
